resolve bug for insert bigdata exception: unsigned short for length

+++ src/ripple/json/json_value.h
@@ -429,7 +429,7 @@ private:
         char* string_;
         ObjectValues* map_{nullptr};
     } value_;
-    unsigned short length_ = 0;
+    unsigned int length_ = 0;
     ValueType type_ : 8;
     int allocated_ : 1;  // Notes: if declared as bool, bitfield is useless.
 };
